Former-commit-id: 6a7db40e72c8d347d6bf20b72604132559e8b8b3
Former-commit-id: 9a597237625a6424d2e81e48f6cc09afaeed0f92

+++ server/src/main/java/org/icgc/dcc/validation/visitor/UniqueFieldsPlanningVisitor.java
@@ -91,7 +91,7 @@ public class UniqueFieldsPlanningVisitor extends InternalFlowPlanningVisitor {
         if(i.hasNext()) {
           TupleEntry firstTuple = i.next();
           bufferCall.getOutputCollector().add(firstTuple.getTupleCopy());
-          int firstOffset = ValidationFields.state(firstTuple).getOffset();
+          long firstOffset = ValidationFields.state(firstTuple).getOffset();
 
           while(i.hasNext()) {
             TupleEntry tupleEntry = i.next();
